complete scheme updation and show times join

+++ backend/src/main/java/com/growkaro/backend/DRO/ApproveUserScheme.java
@@ -1,7 +1,8 @@
 package com.growkaro.backend.DRO;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record ApproveUserScheme(String userSchemeId, Long paidAmount, LocalDate paidDate) {
+public record ApproveUserScheme(String userSchemeId, BigDecimal paidAmount, LocalDate paidDate) {
 
 }
